+++ src/components/UserInfo.jsx
@@ -1,4 +1,4 @@
-import * as Styled from '../pages/MyInfo/styled';
+import styled from 'styled-components';
 import { useNavigate } from 'react-router-dom';
 import { isLoginStorage } from '../utils/loginStorage';
 import UserAccount from './UserAccount';
@@ -52,15 +52,15 @@ const UserInfo = ({ my }) => {
     isLogin ? navigate('/myposting') : navigate('/login');
   };
   return (
-    <Styled.Container>
-      <Styled.InfoWrapper>
-        <Styled.InfoTitle>내 정보</Styled.InfoTitle>
-      </Styled.InfoWrapper>
+    <Container>
+      <InfoWrapper>
+        <InfoTitle>내 정보</InfoTitle>
+      </InfoWrapper>
 
-      <Styled.Wrapper id="top">
-        <Styled.Button id="mobile" onClick={handleNavigate} background="#336af8">
+      <Wrapper id="top">
+        <Button id="mobile" onClick={handleNavigate} background="#336af8">
           {isLogin ? '내가 쓴 글' : '로그인하기'}
-        </Styled.Button>
+        </Button>
         {!isLogin ? (
           <UserAccount
             isLogin={isLogin}
@@ -71,75 +71,75 @@ const UserInfo = ({ my }) => {
           <UserAccount isLogin={isLogin} loginId={my.loginId} email={my.email} />
         )}
 
-        <Styled.Button id="pc" onClick={handleNavigate} background="#336af8">
+        <Button id="pc" onClick={handleNavigate} background="#336af8">
           {isLogin ? '내가 쓴 글' : '로그인하기'}
-        </Styled.Button>
-      </Styled.Wrapper>
+        </Button>
+      </Wrapper>
 
-      <Styled.Wrapper>
+      <Wrapper>
         {!isLogin ? (
           <UserPoint my={fakeUserInfo} isLogin={isLogin} />
         ) : (
           <UserPoint my={my} isLogin={isLogin} />
         )}
-      </Styled.Wrapper>
+      </Wrapper>
 
-      <Styled.Wrapper>
-        <Styled.Content>
-          <Styled.Title>포인트 제도 안내</Styled.Title>
-          <Styled.OptionTitle id="text">
+      <Wrapper>
+        <Content>
+          <Title>포인트 제도 안내</Title>
+          <OptionTitle id="text">
             - 포인트는 강의평가 서비스 내에서만 이용되는 제도입니다.
             <br />- 다른 계정으로 학교 인증 시 해당 계정으로 포인트가 합산 되며, 기존 계정의
             포인트는 초기화 됩니다.
             <br />- 포인트 획득을 위해 허위/중복/성의없는 정보를 작성할 경우, 서비스 이용이 영구
             제한될 수 있습니다.
             <br />- 허위 신고를 남용하는 이용자 또한 제재가 가해질 수 있습니다.
-          </Styled.OptionTitle>
+          </OptionTitle>
           <br />
           <br />
 
-          <Styled.Title>포인트 획득</Styled.Title>
-          <Styled.FlexPointContainer>
+          <Title>포인트 획득</Title>
+          <FlexPointContainer>
             -강의평 작성:
-            <Styled.Color id="pp">+10P</Styled.Color>
-          </Styled.FlexPointContainer>
-          <Styled.FlexPointContainer>
+            <Color id="pp">+10P</Color>
+          </FlexPointContainer>
+          <FlexPointContainer>
             -시험정보 공유:
-            <Styled.Color id="pp">+20P</Styled.Color>
-          </Styled.FlexPointContainer>
-          <Styled.FlexPointContainer>
+            <Color id="pp">+20P</Color>
+          </FlexPointContainer>
+          <FlexPointContainer>
             -신고 보상:
-            <Styled.Color id="pp">+1P</Styled.Color>
-          </Styled.FlexPointContainer>
+            <Color id="pp">+1P</Color>
+          </FlexPointContainer>
 
           <br />
           <br />
-          <Styled.Title>포인트 차감</Styled.Title>
-          <Styled.FlexPointContainer>
+          <Title>포인트 차감</Title>
+          <FlexPointContainer>
             -시험정보 조회:
-            <Styled.Color id="ppp">-20P</Styled.Color>
-          </Styled.FlexPointContainer>
-          <Styled.FlexPointContainer>
+            <Color id="ppp">-20P</Color>
+          </FlexPointContainer>
+          <FlexPointContainer>
             -강의평가/시험정보 삭제:
-            <Styled.Color id="ppp">-30P</Styled.Color>
-          </Styled.FlexPointContainer>
-        </Styled.Content>
-      </Styled.Wrapper>
-      <Styled.Wrapper>
-        <Styled.Content>
-          <Styled.Title>이용 안내</Styled.Title>
+            <Color id="ppp">-30P</Color>
+          </FlexPointContainer>
+        </Content>
+      </Wrapper>
+      <Wrapper>
+        <Content>
+          <Title>이용 안내</Title>
           {isLogin
             ? option.slice(0, 2).map((i) => (
-                <Styled.FlexContainer id="use" key={i.title}>
-                  <Styled.FlexContainer id="last" onClick={() => navigate(`/${i.page}`)}>
+                <FlexContainer id="use" key={i.title}>
+                  <FlexContainer id="last" onClick={() => navigate(`/${i.page}`)}>
                     {i.title}
-                  </Styled.FlexContainer>
-                </Styled.FlexContainer>
+                  </FlexContainer>
+                </FlexContainer>
               ))
             : null}
           {urlOption.map((i) => (
-            <Styled.FlexContainer id="use" key={i.title}>
-              <Styled.FlexContainer
+            <FlexContainer id="use" key={i.title}>
+              <FlexContainer
                 id="last"
                 onClick={() => {
                   i.page === 'email'
@@ -148,22 +148,200 @@ const UserInfo = ({ my }) => {
                 }}
               >
                 {i.title}
-              </Styled.FlexContainer>
-            </Styled.FlexContainer>
+              </FlexContainer>
+            </FlexContainer>
           ))}
           {isLogin
             ? option.slice(2, 4).map((i) => (
-                <Styled.FlexContainer id="use" key={i.title}>
-                  <Styled.FlexContainer id="last" onClick={() => navigate(`/${i.page}`)}>
+                <FlexContainer id="use" key={i.title}>
+                  <FlexContainer id="last" onClick={() => navigate(`/${i.page}`)}>
                     {i.title}
-                  </Styled.FlexContainer>
-                </Styled.FlexContainer>
+                  </FlexContainer>
+                </FlexContainer>
               ))
             : null}
-        </Styled.Content>
-      </Styled.Wrapper>
-    </Styled.Container>
+        </Content>
+      </Wrapper>
+    </Container>
   );
 };
 
 export default UserInfo;
+
+const Wrapper = styled.div`
+  width: 100%;
+  margin: 0 auto;
+
+  &#top {
+    display: flex;
+    @media screen and (max-width: 550px) {
+      flex-direction: column;
+    }
+  }
+`;
+const Container = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 60%;
+  margin: 0 auto;
+
+  font-weight: 400;
+  @media only screen and (max-width: 960px) {
+    width: 90%;
+  }
+`;
+const Content = styled.div`
+  border-radius: 10px;
+  border: 1px solid rgb(224, 224, 224);
+  padding: 1rem 1.8rem;
+  margin: 10px 0;
+  position: relative;
+  &#top {
+    width: 100%;
+    margin-right: 2rem;
+  }
+`;
+const Title = styled.div`
+  font-size: 1.2rem;
+  padding-bottom: 1rem;
+
+  font-weight: 600;
+
+  &#top {
+    padding-bottom: 2rem;
+  }
+`;
+
+const OptionTitle = styled.div`
+  display: flex;
+  width: 160px;
+  @media screen and (max-width: 550px) {
+    font-size: 0.9rem;
+  }
+
+  &#my {
+    color: #222222;
+  }
+
+  &#text {
+    width: 100%;
+    margin-bottom: 10px;
+  }
+
+  &#points {
+    width: 170px;
+  }
+`;
+
+const FlexContainer = styled.div`
+  display: flex;
+  justify-content: space-between;
+  color: #a3a3a3;
+  padding-bottom: 0.5rem;
+
+  @media screen and (max-width: 550px) {
+    font-size: 0.9rem;
+  }
+  &#use {
+    padding: 6px 0;
+  }
+  &#needLogin {
+    color: #222222;
+    position: absolute;
+    top: 45%;
+    left: 40%;
+    text-align: center;
+    line-height: 1.3;
+    font-weight: 500;
+    @media screen and (max-width: 550px) {
+      left: 30%;
+    }
+  }
+  &#col {
+    display: flex;
+    justify-content: center;
+    @media screen and (max-width: 550px) {
+      width: 90%;
+    }
+  }
+  &#last {
+    color: #222222;
+    align-items: center;
+    &:hover {
+      cursor: pointer;
+    }
+  }
+`;
+
+const FlexPointContainer = styled.div`
+  display: flex;
+  color: #515151;
+  padding-bottom: 0.5rem;
+`;
+
+const Button = styled.button`
+  margin: 10px 0;
+  width: 30%;
+  padding: 0 1rem;
+  border: none;
+  background: ${(props) => props.background};
+  color: white;
+  text-align: center;
+  font-size: 1.5rem;
+
+  font-weight: 600;
+  border-radius: 12px;
+  cursor: pointer;
+  user-select: none;
+  &#mobile {
+    display: none;
+  }
+  @media only screen and (max-width: 960px) {
+    font-size: 1.3rem;
+  }
+  @media only screen and (max-width: 550px) {
+    margin: 10px 0;
+    width: 100%;
+    height: 48px;
+    &#mobile {
+      display: block;
+    }
+    &#pc {
+      display: none;
+    }
+  }
+`;
+
+const InfoWrapper = styled.div`
+  display: flex;
+  width: 100%;
+  flex-direction: column;
+  padding: 1rem 0;
+`;
+
+const InfoTitle = styled.div`
+  display: flex;
+  font-size: 1.5rem;
+
+  font-weight: 600;
+  padding-top: 4rem;
+  padding-bottom: 1rem;
+`;
+
+const Color = styled.span`
+  color: #336af8;
+
+  font-size: 1.2rem;
+  @media screen and (max-width: 550px) {
+    font-size: 0.9rem;
+  }
+
+  &#p {
+    color: #6200ee;
+  }
+
+  &#ppp {
+    color: #6200ee;
+  }
+`;
