+++ src/pages/EmailSignUp.jsx
@@ -1,0 +1,71 @@
+import { useLocation, useNavigate } from 'react-router-dom';
+import Button from '../components/Button';
+import Meta from '../components/Meta';
+import { Positioner } from '../components/Wrapper';
+import styled from 'styled-components';
+
+const EmailSignUp = () => {
+  const navigate = useNavigate();
+  const { state } = useLocation();
+  return (
+    <Positioner>
+      <Meta title="SUWIKI : 이메일 인증" />
+      <Wrapper>
+        <Title>이메일 인증</Title>
+        <Img>
+          <img src="images/email.png" alt="options" width={100} />
+        </Img>
+        <Content>
+          <Color>{state}</Color>로<br />
+          전송된 인증 메일을 확인해주세요
+          <br />
+          <br />
+          <br />
+          메일 인증 후 수위키 서비스 이용이 가능합니다.
+        </Content>
+
+        <Button color="blue" onClick={() => navigate('/')}>
+          메인으로
+        </Button>
+      </Wrapper>
+    </Positioner>
+  );
+};
+
+export default EmailSignUp;
+
+export const Title = styled.div`
+  font-size: 4vh;
+  font-weight: 1000;
+  margin: 3rem 0;
+  text-align: center;
+
+  font-weight: 700;
+`;
+
+export const Img = styled.div`
+  text-align: center;
+`;
+
+export const Content = styled.div`
+  text-align: center;
+  margin: 4vh 0;
+  font-weight: bold;
+`;
+
+export const Color = styled.span`
+  color: #4b10f2;
+`;
+
+export const Wrapper = styled.div`
+  margin: 0 auto;
+  @media all and (min-width: 1024px) {
+    width: 500px;
+  }
+  @media all and (min-width: 768px) and (max-width: 1023px) {
+    width: 400px;
+  }
+  @media all and (max-width: 767px) {
+    width: 300px;
+  }
+`;
