--- src/pages/LectureInfo/styled.js
@@ -1,136 +1,0 @@
-import styled, { createGlobalStyle } from 'styled-components';
-
-export const GlobalStyle = createGlobalStyle`
-  #root>div {
-      width: 90%;
-      @media only screen and (min-width: 768px) {
-        width: 90%;
-    }
-  }
-`;
-
-export const Container = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 60%;
-  margin: 0 auto;
-
-  @media screen and (max-width: 960px) {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    width: 90%;
-    margin: 0 auto;
-  }
-`;
-
-export const SearchWrapper = styled.div`
-  display: flex;
-  width: 100%;
-  flex-direction: column;
-  align-items: center;
-`;
-
-export const SearchTitle = styled.div`
-  display: flex;
-  font-size: 1.5rem;
-  margin-right: auto;
-
-  font-weight: 600;
-  padding-top: 4rem;
-`;
-
-export const SearchInput = styled.input`
-  width: 100%;
-  height: 50px;
-  border: none;
-  border-bottom: 1.5px solid #000000;
-  margin: 1.5rem 0;
-  background-image: url('images/icon_search_24.svg');
-  background-repeat: no-repeat;
-  background-position: 99%;
-
-  font-weight: 400;
-  cursor: pointer;
-
-  &:focus {
-    outline: none;
-  }
-
-  @media screen and (max-width: 550px) {
-    font-size: 16px;
-    width: 123%;
-    transform: scale(0.8);
-  }
-`;
-
-export const Content = styled.div`
-  border-radius: 10px;
-  margin: 10px 0;
-  margin-bottom: 3rem;
-
-  &#top {
-    padding: 1rem 2rem;
-    border: 1px solid rgb(224, 224, 224);
-  }
-
-  @media screen and (max-width: 550px) {
-    width: 100%;
-  }
-`;
-
-export const TitleWrapper = styled.div`
-  display: flex;
-  margin-bottom: 0.5rem;
-  align-items: flex-end;
-
-  &#top {
-    justify-content: space-between;
-    align-items: flex-start;
-  }
-  &#bottom {
-    margin-top: 1rem;
-  }
-`;
-
-export const Wrapper = styled.div`
-  width: 100%;
-  margin: 0 auto;
-`;
-
-export const MenuTitle = styled.li`
-  font-size: 16px;
-  margin-bottom: 1rem;
-  text-align: center;
-
-  display: flex;
-  color: lightgray;
-  padding-right: 1rem;
-
-  &:hover {
-    cursor: pointer;
-  }
-  &#${({ check }) => check} {
-    color: black;
-  }
-`;
-
-export const Writing = styled.img`
-  &:hover {
-    cursor: pointer;
-  }
-`;
-
-export const FlexContainer = styled.div`
-  display: flex;
-  color: #a3a3a3;
-  width: 45%;
-  &#col {
-    flex-direction: column;
-  }
-
-  @media screen and (max-width: 550px) {
-    width: 100%;
-  }
-`;
