+++ src/pages/Main.jsx
@@ -1,3 +1,12 @@
+import { useEffect, useState } from 'react';
+import Modal from 'react-modal';
+import { useNavigate } from 'react-router-dom';
+import Major from '../api/Major';
+import { versionCheck } from '../app/versionCheck';
+import MainList from '../components/MainList';
+import MajorSearch from '../components/MajorSearch';
+import { MajorModalStyle } from '../components/ModalStyle';
+import { majorList } from './Search';
 import styled, { createGlobalStyle } from 'styled-components';
 import * as React from 'react';
 import PropTypes from 'prop-types';
@@ -5,6 +14,141 @@ import SelectUnstyled, { selectUnstyledClasses } from '@mui/base/SelectUnstyled'
 import OptionUnstyled, { optionUnstyledClasses } from '@mui/base/OptionUnstyled';
 import { styled as styled_mui } from '@mui/system';
 import PopperUnstyled from '@mui/base/PopperUnstyled';
+
+const options = [
+  {
+    name: '최근 올라온 강의',
+    lec: 'modifiedDate',
+    imgs: 'images/icon_color_fire_36.svg',
+  },
+  {
+    name: '꿀 강의',
+    lec: 'lectureHoneyAvg',
+    imgs: 'images/icon_color_bee_36.svg',
+  },
+  {
+    name: '만족도가 높은 강의',
+    lec: 'lectureSatisfactionAvg',
+    imgs: 'images/icon_color_thumbs_36.svg',
+  },
+  {
+    name: '배울게 많은 강의',
+    lec: 'lectureLearningAvg',
+    imgs: 'images/icon_color_book_36.svg',
+  },
+  {
+    name: 'BEST 강의',
+    lec: 'lectureTotalAvg',
+    imgs: 'images/icon_color_best_36.svg',
+  },
+];
+const Main = () => {
+  const navigate = useNavigate();
+  const [search, setSearch] = useState('');
+  const [lecture, setLecture] = useState('modifiedDate');
+  const [checkClass, setCheckClass] = useState('전체');
+  const [modalIsOpen, setModalIsOpen] = useState(false);
+  const major = Major();
+  const onChange = (e) => {
+    setSearch(e.currentTarget.value);
+  };
+
+  const onChangeHandler = (e) => {
+    setLecture(e);
+  };
+
+  const onKeypress = (e) => {
+    if (e.key === 'Enter') {
+      if (e.currentTarget.value.length < 2) {
+        alert('두 글자 이상 입력해주세요');
+      } else {
+        navigate(`/search?q=${search}&option=lectureTotalAvg&majorType=전체`);
+      }
+    }
+  };
+  useEffect(() => {
+    versionCheck(major);
+  }, [major]);
+  return (
+    <div>
+      <Banner>
+        <BannerWrapper>
+          <div>
+            수위키,
+            <br />
+            강의평가의 모든 것<BannerSub>강의평가의 모든 것 수위키</BannerSub>
+          </div>
+          <picture>
+            <source srcSet="/images/resize_banner.avif" type="image/avif" />
+            <source srcSet="/images/resize_banner.webp" type="image/webp" />
+            <source srcSet="/images/banner.webp" type="image/webp" />
+            <source srcSet="/images/banner.svg" type="image/png" />
+            <BannerImg src="images/banner.svg" alt="banner" width={450} height={450} />
+          </picture>
+        </BannerWrapper>
+      </Banner>
+      <Container>
+        <SearchWrapper>
+          <SearchTitle>강의평가 검색</SearchTitle>
+          <SearchInput
+            onChange={onChange}
+            placeholder="강의명, 교수명으로 원하는 강의평가를 찾아보세요"
+            onKeyPress={onKeypress}
+            onMouseLeave={() => document.activeElement.blur()}
+          />
+        </SearchWrapper>
+        <SearchWrapper>
+          <HeadSelection>
+            <FlexWrapper onClick={() => setModalIsOpen(true)}>
+              <SortSelect id="major" defaultValue={checkClass}>
+                {majorList.map((index) => (
+                  <StyledOption id="semester" key={index} value={index}>
+                    <Soption id="semester">{checkClass}</Soption>
+                  </StyledOption>
+                ))}
+              </SortSelect>
+            </FlexWrapper>
+            <CustomSelect defaultValue={'modifiedDate'} onChange={onChangeHandler}>
+              {options.map((index) => (
+                <StyledOption key={index.name} value={index.lec}>
+                  <Soption>
+                    <Img alt="option-icon" width="22" height="22" src={index.imgs} /> {index.name}
+                  </Soption>
+                </StyledOption>
+              ))}
+            </CustomSelect>
+          </HeadSelection>
+          <HeadSelection>
+            <MainList lecture={lecture} checkClass={checkClass} />
+          </HeadSelection>
+        </SearchWrapper>
+        <Button
+          background="#336af8"
+          onClick={() => {
+            navigate(`/search?q=&option=lectureTotalAvg&majorType=전체`);
+          }}
+        >
+          더 보러 가기 →
+        </Button>
+      </Container>
+      <Modal
+        isOpen={modalIsOpen}
+        style={MajorModalStyle}
+        // 오버레이나 esc를 누르면 핸들러 동작
+        ariaHideApp={false}
+        onRequestClose={() => setModalIsOpen(false)}
+      >
+        <MajorSearch
+          checkClass={checkClass}
+          setModalIsOpen={setModalIsOpen}
+          setCheckClass={setCheckClass}
+        />
+      </Modal>
+    </div>
+  );
+};
+
+export default Main;
 
 export const GlobalStyle = createGlobalStyle`
   #root>div {
