+++ src/pages/Notice.jsx
@@ -2,20 +2,20 @@ import { Fragment, useEffect } from 'react';
 import { useInView } from 'react-intersection-observer';
 import { useInfiniteQuery } from 'react-query';
 import { useNavigate } from 'react-router-dom';
-import Notices from '../../api/Notice';
-import Meta from '../../components/Meta';
-import Spinner from '../../components/Spinner';
-import * as Styled from './styled';
+import Notices from '../api/Notice';
+import Meta from '../components/Meta';
+import Spinner from '../components/Spinner';
+import styled from 'styled-components';
 
 export const NoticeItem = ({ notice }) => {
   const navigate = useNavigate();
   const onClick = () => navigate(`/notice/detail?id=${notice.id}`);
 
   return (
-    <Styled.NoticeWrap onClick={onClick}>
-      <Styled.Title>{notice.title}</Styled.Title>
-      <Styled.Option>{notice.modifiedDate.slice(0, 10)}</Styled.Option>
-    </Styled.NoticeWrap>
+    <NoticeWrap onClick={onClick}>
+      <Title>{notice.title}</Title>
+      <Option>{notice.modifiedDate.slice(0, 10)}</Option>
+    </NoticeWrap>
   );
 };
 
@@ -55,7 +55,7 @@ export const NoticeContainer = () => {
           </div>
         </>
       ) : (
-        <Styled.NoNotice>아직 공지사항이 없어요.</Styled.NoNotice>
+        <NoNotice>아직 공지사항이 없어요.</NoNotice>
       )}
     </>
   );
@@ -63,12 +63,72 @@ export const NoticeContainer = () => {
 
 const Notice = () => {
   return (
-    <Styled.AppContainer>
+    <AppContainer>
       <Meta title="SUWIKI : 공지사항" />
-      <Styled.AppTitle>공지사항</Styled.AppTitle>
+      <AppTitle>공지사항</AppTitle>
       <NoticeContainer />
-    </Styled.AppContainer>
+    </AppContainer>
   );
 };
 
 export default Notice;
+
+const AppContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 60%;
+  margin: 0 auto;
+  padding-bottom: 5rem;
+
+  @media screen and (max-width: 960px) {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    width: 90%;
+    margin: 0 auto;
+  }
+`;
+
+const AppTitle = styled.div`
+  display: flex;
+  width: 100%;
+  font-size: 1.5rem;
+
+  font-weight: 600;
+  padding-top: 5rem;
+  padding-bottom: 1rem;
+`;
+
+const Title = styled.div`
+  display: flex;
+  font-size: 1rem;
+  margin-right: 0.7rem;
+`;
+
+const Option = styled.div`
+  border-radius: 10px;
+  padding: 0.5rem 0;
+  font-size: 0.8rem;
+  color: #a3a3a3;
+
+  font-weight: 300;
+`;
+
+const NoticeWrap = styled.div`
+  width: 100%;
+  border: 1.5px solid #f1f1f1;
+  padding: 1.5rem 1.5rem;
+  border-radius: 10px;
+  margin-top: 2vh;
+
+  &:hover {
+    cursor: pointer;
+  }
+`;
+
+const NoNotice = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+`;
