--- src/pages/Search/styled.js
@@ -1,92 +1,0 @@
-import styled, { createGlobalStyle } from 'styled-components';
-
-export const GlobalStyle = createGlobalStyle`
-  #root>div {
-      width: 90%;
-      @media only screen and (min-width: 768px) {
-        width: 90%;
-    }
-  }
-`;
-
-export const Container = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 60%;
-  margin: 0 auto;
-  @media only screen and (max-width: 960px) {
-    width: 90%;
-  }
-`;
-
-export const SearchResultWrapper = styled.div`
-  display: flex;
-  width: 100%;
-  margin-top: 0.5rem;
-  justify-content: space-between;
-  align-items: center;
-`;
-
-export const SearchWrapper = styled.div`
-  display: flex;
-  width: 100%;
-  flex-direction: column;
-  align-items: center;
-  padding: 1rem 0;
-`;
-
-export const SearchTitle = styled.div`
-  display: flex;
-  font-size: 1.5rem;
-  margin-right: auto;
-
-  font-weight: 600;
-  padding-top: 4rem;
-`;
-
-export const SearchInput = styled.input`
-  width: 100%;
-  height: 50px;
-  border: none;
-  border-bottom: 1.5px solid #000000;
-  margin: 1.5rem 0;
-  background-image: url('images/icon_search_24.svg');
-  background-repeat: no-repeat;
-  background-position: 99%;
-
-  font-weight: 400;
-  cursor: pointer;
-
-  &:focus {
-    outline: none;
-  }
-
-  @media screen and (max-width: 550px) {
-    font-size: 16px;
-    width: 123%;
-    transform: scale(0.8);
-  }
-`;
-
-export const HeadSelection = styled.div`
-  padding-bottom: 10px;
-  width: 100%;
-`;
-
-export const FlexWrapper = styled.div`
-  display: flex;
-  margin-bottom: 1rem;
-
-  &#count {
-    @media screen and (max-width: 960px) {
-      display: none;
-    }
-  }
-`;
-
-export const Color = styled.span`
-  color: #336af8;
-  margin-left: 6px;
-  text-align: center;
-`;
